# lib/api-zod/src/generated/types/analyzeFeedbackBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeFeedbackBody {
  /** The raw user feedback text to analyze */
  feedback: string;
  /** Optional product context to help Claude understand the domain */
  productContext?: string;
}