# lib/api-zod/src/generated/types/feedbackAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackSummary } from "./feedbackSummary";
import type { FeedbackTheme } from "./feedbackTheme";

export interface FeedbackAnalysis {
  themes: FeedbackTheme[];
  summary: FeedbackSummary;
}