# lib/api-zod/src/generated/types/feedbackSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackSummaryOverallSentiment } from "./feedbackSummaryOverallSentiment";

export interface FeedbackSummary {
  overallSentiment: FeedbackSummaryOverallSentiment;
  /** Sentiment score from -1 (very negative) to 1 (very positive) */
  sentimentScore: number;
  totalThemes: number;
  criticalIssues: number;
  keyStrengths: string[];
  keyWeaknesses: string[];
  recommendedActions: string[];
  executiveSummary: string;
}