# lib/api-zod/src/generated/types/feedbackTheme.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackThemePriority } from "./feedbackThemePriority";
import type { FeedbackThemeSentiment } from "./feedbackThemeSentiment";

export interface FeedbackTheme {
  name: string;
  description: string;
  occurrences: number;
  quotes: string[];
  sentiment: FeedbackThemeSentiment;
  priority: FeedbackThemePriority;
}